Use Num volume in ExtendedHistoricalQuote constructors

+++ timeseries-stockfeed/src/main/java/com/leonarduk/finance/stockfeed/datatransformation/correction/BadScalingCorrector.java
@@ -9,7 +9,6 @@ import org.ta4j.core.num.Num;
 import java.time.ZoneId;
 import java.util.Iterator;
 import java.util.List;
-import java.time.ZoneId;
 
 public class BadScalingCorrector implements TimeSeriesCleaner {
 
@@ -39,7 +38,7 @@ public class BadScalingCorrector implements TimeSeriesCleaner {
                         scaleDown(current.getHighPrice()),
                         scaleDown(current.getClosePrice()),
 
-                        current.getVolume().longValue(),
+                        current.getVolume(),
                         comment + ": Scaled from " + current.getClosePrice() + " to " + scaleDown(current.getClosePrice()));
                 cleanedSeries.add(cleanedQuote);
             } else if (current.getClosePrice().multipliedBy(DoubleNum.valueOf(SCALE))
@@ -52,7 +51,7 @@ public class BadScalingCorrector implements TimeSeriesCleaner {
                         scaleUp(current.getHighPrice()),
                         scaleUp(current.getClosePrice()),
 
-                        current.getVolume().longValue(),
+                        current.getVolume(),
                         "Scaled from " + current.getClosePrice() + " to " + scaleUp(current.getClosePrice()));
                 cleanedSeries.add(cleanedQuote);
             } else {
